Refactor ticker management in Hermes dashboard
- Removed legacy redirect pages for ticker editing, consolidating functionality into modals.
- Introduced a new TickerEditForm component for editing ticker details, enhancing user experience.
- Updated TickerEditModal to utilize the new form structure, streamlining the editing process.
